chore: renamed rentalsservice
