+++ src/routes.ts
@@ -2,7 +2,7 @@ import { Application } from 'express'
 
 import { routes as tenantRoutes } from './services/tenantservice'
 import { routes as leaseContactRoutes } from './services/leasecontractservice'
-import { routes as rentalRoutes } from './services/rentalsservice'
+import { routes as rentalRoutes } from './services/rentalservice'
 import { routes as healthRoutes } from './health'
 
 export { tenantRoutes, leaseContactRoutes, rentalRoutes, healthRoutes }
