+++ src/services/rentalservice/__fixtures__/fastAPIAdapterResult.fixture.ts
